Separation of the system and benchmark into separate modules, adding tests on the number of messages sent between agents
