+++ benchmark/Dockerfile
@@ -11,4 +11,4 @@ RUN yes | apt install maven
 
 RUN mvn clean install
 
-CMD cd ./messaging; java -cp "../jade.jar:./target/messaging-1.0-SNAPSHOT.jar" jade.Boot -agents "CounterAgent:com.jade.CounterAgent;SenderAgent:com.jade.SenderAgent"
+CMD cd ./messaging; java -cp "../jade.jar:./target/messaging-1.0-SNAPSHOT.jar" jade.Boot -agents "CounterAgent:com.jade.system.CounterAgent;SenderAgent:com.jade.system.SenderAgent"
