+++ benchmark/hello/src/main/java/com/jade/benchmark/HelloWorldAgentTest.java
@@ -1,32 +1,22 @@
-package com.jade;
+package com.jade.benchmark;
 
+import com.jade.system.HelloWorldAgent;
 import jade.core.Profile;
 import jade.core.ProfileImpl;
 import jade.core.Runtime;
 import jade.wrapper.AgentContainer;
 import jade.wrapper.AgentController;
 import jade.wrapper.StaleProxyException;
-import org.junit.Test;
-import org.openjdk.jmh.annotations.Benchmark;
-import org.openjdk.jmh.annotations.BenchmarkMode;
-import org.openjdk.jmh.annotations.Fork;
-import org.openjdk.jmh.annotations.Measurement;
-import org.openjdk.jmh.annotations.Mode;
-import org.openjdk.jmh.annotations.OutputTimeUnit;
-import org.openjdk.jmh.annotations.Param;
-import org.openjdk.jmh.annotations.Scope;
-import org.openjdk.jmh.annotations.State;
+import org.openjdk.jmh.annotations.*;
 
 import java.util.concurrent.TimeUnit;
-
-import static org.junit.Assert.fail;
 
 @State(Scope.Thread)
 public class HelloWorldAgentTest {
 
     private AgentContainer container;
 
-//    @Param({"1", "2", "5", "10", "50", "100"})
+//    @Param({"1", "2", "5", "10", "50", "100", "1000", "10000}) -- If you have strong PC
     @Param({"1", "3"})
     private int numberOfAgentsCreated;
 
@@ -49,7 +39,6 @@ public class HelloWorldAgentTest {
     @Measurement(iterations = 2)
     @Fork(value = 4)
     @OutputTimeUnit(TimeUnit.SECONDS)
-    @Test
     public void testHelloWorldAgent() {
         // Before
         setup();
@@ -68,7 +57,6 @@ public class HelloWorldAgentTest {
             agentController.start();
         } catch (StaleProxyException e) {
             e.printStackTrace();
-            fail("Test failed: " + e.getMessage());
         }
     }
 }
